% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_cambio_contrasena.R
\name{actualizar_contrasena_usuario}
\alias{actualizar_contrasena_usuario}
\title{Registrar nueva contrase침a}
\usage{
actualizar_contrasena_usuario(conexion, fecha_hora_evento, usuario, contrasena)
}
\arguments{
\item{conexion}{clase formal. Conexi칩n base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}

\item{contrasena}{clase character}
}
\description{
Esta funci칩n registra la nueva contrase침a de un usuario en la tabla ''
}
