% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_miembros.R
\name{actualizar_control_modificacion_datos_miembros}
\alias{actualizar_control_modificacion_datos_miembros}
\title{Actualizar control de modificación datos miembros}
\usage{
actualizar_control_modificacion_datos_miembros(datos)
}
\arguments{
\item{datos}{clase dataframe. Datos miembros}
}
\description{
Esta función actualiza el control de modificación datos miembros
}
