% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{actualizar_control_modificacion_datos_usuarios}
\alias{actualizar_control_modificacion_datos_usuarios}
\title{Actualizar control de modificación datos usuarios}
\usage{
actualizar_control_modificacion_datos_usuarios(datos)
}
\arguments{
\item{datos}{clase dataframe. Datos usuarios}
}
\description{
Esta función actualiza el control de modificación datos usuarios
}
