% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_general.R
\name{agregar_botones_tabla}
\alias{agregar_botones_tabla}
\title{Agrega columna botones a data.table (DT)}
\usage{
agregar_botones_tabla(llave, ns, botones)
}
\arguments{
\item{llave}{clase character. Llave de los botones}

\item{ns}{clase character. Namespace al que pertenece el boton}

\item{botones}{clase vector character. Vector de los nombres de los botones}
}
\description{
Esta función agrega la columna botones a una tabla tipo data.table (DT)
}
