% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{boton_reactivo}
\alias{boton_reactivo}
\title{Crea boton reactivo}
\usage{
boton_reactivo(
  botones = NULL,
  inputId = NULL,
  icon = "plus",
  status = "default"
)
}
\arguments{
\item{botones}{clase lista. Lista de botones}

\item{inputId}{clase character. Id del boton}

\item{icon}{clase character. Icono del boton}

\item{status}{clase character. Status del boton}
}
\description{
Esta función crea el boton reactivo de una lista de botones
}
