% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{botones_generales}
\alias{botones_generales}
\title{Crear lista de botones generales}
\usage{
botones_generales(modulo_activo = NULL, administrador = FALSE)
}
\arguments{
\item{modulo_activo}{clase character}

\item{administrador}{clase boolean}
}
\description{
Esta función crea la lista de los botones generales con base en el modulo activo
y en si el usuario es administrador
}
