% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_inicio_sesion.R
\name{buscar_permisos_perfil}
\alias{buscar_permisos_perfil}
\title{Buscar permisos perfil}
\usage{
buscar_permisos_perfil(conexion, perfil)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{perfil}{clase character}
}
\description{
Esta función busca y extra los permisos de un perfil en la tabla ''
}
