% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_inicio_sesion.R
\name{buscar_usuario}
\alias{buscar_usuario}
\title{Buscar usuario}
\usage{
buscar_usuario(conexion, usuario)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{usuario}{clase character}
}
\description{
Esta función busca y extra la información de un usuario en la tabla ''
}
