% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{colapsar_columnas_dataframe}
\alias{colapsar_columnas_dataframe}
\title{Colapsar columnas data.frame en una cadena de texto}
\usage{
colapsar_columnas_dataframe(datos)
}
\arguments{
\item{datos}{clase data.frame}
}
\description{
Esta función colapsa las columnas de un data.frame en una cadena de texto
}
