% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_general.R
\name{contenedor}
\alias{contenedor}
\title{Agrega contenido html a un contenedor}
\usage{
contenedor(contenido)
}
\arguments{
\item{contenido}{clase html. Contenido del contenerdor}
}
\description{
Esta función agrega el contenido html a un contenerdor (wellPanel)
}
