% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{desencriptar_contrasena}
\alias{desencriptar_contrasena}
\title{Desencriptar contraseña}
\usage{
desencriptar_contrasena(contrasena)
}
\arguments{
\item{contrasena}{clase character}
}
\description{
Esta función desencripta la contraseña por medio de un key y un nonce
}
