% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{elimnar_caracteres_especiales}
\alias{elimnar_caracteres_especiales}
\title{Elimniar caracteres especiales}
\usage{
elimnar_caracteres_especiales(texto)
}
\arguments{
\item{texto}{clase character.}
}
\description{
Esta función elimina los caracteres especiales de una cadena de texto
}
