% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{encriptar_contrasena}
\alias{encriptar_contrasena}
\title{Encriptar contraseña}
\usage{
encriptar_contrasena(contrasena)
}
\arguments{
\item{contrasena}{clase character}
}
\description{
Esta función encripta la contraseña por medio de un key, nonce y la semilla del computador
}
