% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{enviar_correo_bienvenida}
\alias{enviar_correo_bienvenida}
\title{Enviar correo de bienvenida}
\usage{
enviar_correo_bienvenida(usuario, nombre, correo, contrasena)
}
\arguments{
\item{usuario}{clase character}

\item{nombre}{clase character}

\item{correo}{clase character}

\item{contrasena}{clase character}
}
\description{
Esta función envia correo el correo de bienvenida al nuevo usuario
}
