% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_inicio_sesion.R
\name{enviar_correo_codigo_verificacion}
\alias{enviar_correo_codigo_verificacion}
\title{Enviar correo con el código de verificación}
\usage{
enviar_correo_codigo_verificacion(nombre, correo, codigo)
}
\arguments{
\item{nombre}{clase character}

\item{correo}{clase character}

\item{codigo}{clase character}
}
\description{
Esta función envia correo con el código de verificación al usuario
}
