% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{escribir_log}
\alias{escribir_log}
\title{Agregar log}
\usage{
escribir_log(
  logs,
  usuario = NA_character_,
  modulo = NA_character_,
  fecha_hora_evento = NA_character_,
  evento,
  tipo = NA_character_,
  resultado = 1,
  resumen = NA_character_,
  detalle = NA_character_
)
}
\arguments{
\item{logs}{clase data.frame}

\item{usuario}{clase character}

\item{modulo}{clase character}

\item{fecha_hora_evento}{clase character}

\item{evento}{clase character}

\item{tipo}{clase character}

\item{resultado}{clase numeric. 1 si el evento fue exitoso, 0 de lo contrario.}

\item{resumen}{clase character}

\item{detalle}{clase character}
}
\description{
Esta función agrega un nuevo log a data.frame logs
}
