% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{insertar_activar_ui_tabpanels}
\alias{insertar_activar_ui_tabpanels}
\title{Inserta o activa ui tabPanel}
\usage{
insertar_activar_ui_tabpanels(
  session,
  tabpanels_activos = NA,
  tabpanels_nuevos = NA,
  remover = FALSE
)
}
\arguments{
\item{session}{clase session. Variable que almacena la sesión del usuario}

\item{tabpanels_activos}{clase vector character. Vector de los id's de los tabPanels activos}

\item{tabpanels_nuevos}{clase list/vector character. Lista que contiene las ui de los tabPanels que se insertaran o vector
los id's de los tabPanels que se desean activar.}

\item{remover}{clase boolean. TRUE si se desea remover los tabPanels activos o FALSE si se desea
esconder los tabPanels activos Por defecto FALSE}
}
\description{
Esta función inserta o activa la ui de uno o varios tabPanels al navbarPage
}
