% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{insertar_tabpanels}
\alias{insertar_tabpanels}
\title{Inserta ui tabPanel}
\usage{
insertar_tabpanels(session, tabpanels = NA)
}
\arguments{
\item{session}{clase session. Variable que almacena la sesión del usuario}

\item{tabpanels}{clase vector character. Vector de los id's de los tabPanels activos}
}
\description{
Esta función inserta o activa la ui de uno o varios tabPanels al navbarPage
}
