% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_general.R
\name{pie_pagina}
\alias{pie_pagina}
\title{Crear pie de pagina}
\usage{
pie_pagina(posicion = "fixed")
}
\arguments{
\item{posicion}{clase character. Posición del pie de pagina ("fixed", "absolute"). Por defecto "fixed"}
}
\description{
Esta función crea el pie de pagina
}
