% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{registrar_bloqueo_usuario}
\alias{registrar_bloqueo_usuario}
\title{Registrar bloqueo usuario}
\usage{
registrar_bloqueo_usuario(conexion, fecha_hora_evento, usuario)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}
}
\description{
Esta función registra el bloqueo del usuario modificado en la tabla adm_usuarios
}
