% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_inicio_sesion.R
\name{registrar_conexion_exitosa_usuario}
\alias{registrar_conexion_exitosa_usuario}
\title{Registrar conexión exitosa usuario}
\usage{
registrar_conexion_exitosa_usuario(conexion, fecha_hora_evento, datos)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{datos}{clase data.frame. Los datos del usuario deben ser los generados por la función
\code{\link{buscar_usuario}} o tener una estructura igual a dichos datos}
}
\description{
Esta función registra la conexión exitosa de un usuario en la tabla ''
}
