% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_inicio_sesion.R
\name{registrar_conexion_fallida_usuario}
\alias{registrar_conexion_fallida_usuario}
\title{Registrar conexi贸n fallida usuario}
\usage{
registrar_conexion_fallida_usuario(conexion, fecha_hora_evento, usuario)
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}
}
\description{
Esta funci贸n registra la conexi贸n fallida de un usuario en la tabla ''
}
