% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_verificacion_usuario.R
\name{registrar_cookie_usuario}
\alias{registrar_cookie_usuario}
\title{Registrar nueva cookie}
\usage{
registrar_cookie_usuario(conexion, datos)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{datos}{clase data.frame. Los datos de la nueva cookie}
}
\description{
Esta función registra la nueva cookie de un usuario en la tabla ''
}
