% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_miembros.R
\name{registrar_eliminacion_miembro}
\alias{registrar_eliminacion_miembro}
\title{Registrar eliminaci贸n miembro}
\usage{
registrar_eliminacion_miembro(conexion, id)
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{id}{clase character}
}
\description{
Esta funci贸n registra registra la eliminaci贸n del miembro (elimina al miembro) en la tabla adm_miembros
}
