% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{registrar_eliminacion_usuario}
\alias{registrar_eliminacion_usuario}
\title{Registrar eliminaci贸n usuario}
\usage{
registrar_eliminacion_usuario(conexion, usuario)
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{usuario}{clase character}
}
\description{
Esta funci贸n registra registra la eliminaci贸n del usuario (elimina al usuario) en la tabla adm_usuarios
}
