% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_miembros.R
\name{registrar_modificacion_miembro}
\alias{registrar_modificacion_miembro}
\title{Registrar modificaci贸n miembro}
\usage{
registrar_modificacion_miembro(
  conexion,
  fecha_hora_evento,
  id,
  nombre,
  nombre_abreviacion,
  nit,
  tipo,
  clase,
  estado,
  segmentos
)
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}

\item{correo}{clase character}

\item{perfil}{clase character}
}
\description{
Esta funci贸n registra la nueva informaci贸n del miembro modificado en la tabla ADM_MIEMBROS
}
