% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{registrar_modificacion_usuario}
\alias{registrar_modificacion_usuario}
\title{Registrar modificaci贸n usuario}
\usage{
registrar_modificacion_usuario(
  conexion,
  fecha_hora_evento,
  usuario,
  correo,
  perfil
)
}
\arguments{
\item{conexion}{clase formal. Conexi贸n base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}

\item{correo}{clase character}

\item{perfil}{clase character}
}
\description{
Esta funci贸n registra la nueva informaci贸n del usuario modificado en la tabla adm_usuarios
}
