% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{registrar_nueva_contrasena_usuario}
\alias{registrar_nueva_contrasena_usuario}
\title{Registrar nueva contrase침a usuario}
\usage{
registrar_nueva_contrasena_usuario(
  conexion,
  fecha_hora_evento,
  usuario,
  contrasena
)
}
\arguments{
\item{conexion}{clase formal. Conexi칩n base de datos}

\item{fecha_hora_evento}{clase datetime}

\item{usuario}{clase character}

\item{contrasena}{clase character}
}
\description{
Esta funci칩n registra la nueva contrase침a del usuario modificado en la tabla adm_usuarios
}
