% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_miembros.R
\name{registrar_nuevo_miembro}
\alias{registrar_nuevo_miembro}
\title{Registrar nuevo miembro}
\usage{
registrar_nuevo_miembro(conexion, datos)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{datos}{clase data.frame. Los datos del miembro}
}
\description{
Esta función registra un nuevo miembro en la tabla adm_miembros
}
