% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_administrador_usuarios.R
\name{registrar_nuevo_usuario}
\alias{registrar_nuevo_usuario}
\title{Registrar nuevo usuario}
\usage{
registrar_nuevo_usuario(conexion, datos)
}
\arguments{
\item{conexion}{clase formal. Conexión base de datos}

\item{datos}{clase data.frame. Los datos del usuario}
}
\description{
Esta función registra un nuevo usuario en la tabla adm_usuarios
}
