% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{retroalimientacion_evento}
\alias{retroalimientacion_evento}
\title{Retroalimentaci贸n evento}
\usage{
retroalimientacion_evento(
  input,
  output,
  session,
  logs,
  estado_evento,
  boton,
  modal = 1
)
}
\arguments{
\item{input}{clase input. Variable que almacena la input. del usuario}

\item{output}{clase output. Variable que almacena la output del usuario}

\item{session}{clase session. Variable que almacena la sesi贸n del usuario}

\item{logs}{clase data.frame}

\item{estado_evento}{clase list}

\item{boton}{clase character}

\item{modal}{clase numeric}
}
\description{
Esta funci贸n genera la retroalimentaci贸n de un evento, generando el log
y notificandondole al usuario el resultado.
}
