% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{rsc_card}
\alias{rsc_card}
\title{Grilla Tarjetas}
\usage{
rsc_card(content)
}
\arguments{
\item{content}{clase data.frame. Requiere contener las siguientes columnas "guid," "url", "title".
y las siguientes columnsa de forma opcional "owner_username", "description" y "updated_time".}
}
\description{
Renderiza el contenido en una grilla de tarjetas
}
