% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_general.R
\name{tarjeta}
\alias{tarjeta}
\title{Agrega contenido html a una tarjeta}
\usage{
tarjeta(title, width, ...)
}
\arguments{
\item{title}{clase character. Título de la tarjeta}

\item{width}{clase integer. Ancho de la tarjeta minimo 1 y maximo 12}

\item{...}{clase html. Contenido de la tarjeta}
}
\description{
Esta función agrega el contenido html a una tarjeta (card)
}
