% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{verificar_formato_fecha}
\alias{verificar_formato_fecha}
\title{Verifica el formato de una fecha es correcto}
\usage{
verificar_formato_fecha(fecha)
}
\arguments{
\item{fecha}{clase date/character}
}
\description{
Esta función verifica si el formato de una fecha es correcto (YYYY-MM-DD)
}
