% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_general.R
\name{escribir_log}
\alias{escribir_log}
\title{Escribir log}
\usage{
escribir_log(
  logs,
  id = NA_real_,
  usuario = NA_character_,
  perfil = NA_character_,
  aplicacion = NA_character_,
  fecha_hora = NA_character_,
  evento = NA_character_,
  descripcion = NA_character_,
  resultado = 1,
  resumen = NA_character_,
  detalle = NA_character_
)
}
\arguments{
\item{logs}{clase data.frame}

\item{id}{clase number}

\item{usuario}{clase character}

\item{perfil}{clase character}

\item{aplicacion}{clase character}

\item{fecha_hora}{clase character}

\item{evento}{clase character}

\item{descripcion}{clase character}

\item{resultado}{clase numeric. 1 si el evento fue exitoso, 0 de lo contrario.}

\item{resumen}{clase character}

\item{detalle}{clase character}
}
\description{
Esta función ecribe un nuevo log a data.frame logs
}
